app/interface/py/requirements.txt
tornado==6.0.4
Flask==1.1.1
celery==4.4.6
cryptography==2.8
PyMySQL==0.10.0
dbutils==1.3
redis==3.4.1
kafka==1.3.5
sqlalchemy==1.3.18
gunicorn==20.0.4
pytest==5.4.1
requests==2.23.0
gevent==1.4.0
amqp==2.6.0
python-dotenv==0.14.0